Started on reworking to separate the data from the UI
This app was a standalone app, but due having a backend now this app
will now become like any traditional frontend.

All the data is now placed as mockdata for the gulp-rest-emulator.
Added gulp-rest-emulator to emulate a REST backend
Started with renaming ispfield to category to match the naming in the
backend.
Removed Storage.js since the backend will do that.
Removed the header of the page, the backend will run the UI in a
separate window with a header.
Removed the generate isp pdf file since the backend will do that