Added facultyctrl
Fetches the faculties,masters and tracks from the backend.
Changed the behaviour of init for all the controllers.
When FacultyCtrl file is included, it will start fetch.
When masters.loaded is emitted it will start fetching courses
When courses.loaded is emitted it will start fetching categories.

The view does not control when to init the controllers.
When including a controller for the first time it will call init.