Renamed fields to category in all components. Update naming of rest data
All the id are now named to the entity id the represent. So the id in
the list of categories is now catId for example.

Also fixed a bug in the year view header where the quarter ects was not
being calculated properly.

Fixed a bug in switching between track. When switching the categories
were not being updated properly