Another huge update, but this time only for the dev
Changed the build system from gulp/browserify/watchify to webpack.
Also moved away from less to plain css. When there is a need for sass i
will add it to the build chain.

Index.html is now in the www directory indicating that it contains html.
The css/less/sass files can now be imported in the components with
import 'whatever.css' THIS IS NOW THE WAY TO GO.
No global css. The css that is being imported should only effect the
components which is being included for.