CourseCtrl.added consists only of ids, fixed the inital state
In some cases coursetree were added to CourseCtrl.added and in other
cases only coursedetail. So to make it more consistent, it now onlly
consists of ids.

Some inital state of some components were default value like 0 or '' but
the default value should be the calculated value, for example in
CourseTree the ects inital value was 0 but that should be whatever
CourseCtrl.addedEcts returns at that point.

This fixes some inconsisted states when making a component visible or
not