Fixed a bug regarding duplicate courses
The storage was listening to add/remove events but due to the nature of
the order of the events executed it would save an illegal state of the
app. A state that has occured but is illegal.

So the most simple option was just to make Storage.save explicit.
When you want to save the current state, you should call Storage.save.

It now also shows the errors when loading the page.

Fixes #14 fixes #13