Made it possible to drag between sidebar/yearview/categories
You can now drag a course or a group from the sidebar to the yearview or
a category. There is still a ghosting issue. When you drag a course from
the sidebar it dissapears :/

It is not yet possible to drag from the yearview to the sidebar.
I have to abstract the CourseDnD and the CourseTree drag feature for
that first. So it gets easier to drag a course item without caring in
which comp it is being used.

It is possible to move a course from a category to the sidebar, this is
a remove action.