Started working in displaying multiple years
This new feature will allow you to specify which course you are planning
to do in which year.

For now updated the GET /categories/:trackId which now also passes the
years. A year looks like: {year: 2016, courses: [courseIds...]}.

It was the whish of the backender to have everything in one request
path.

When fetching the categories it will show the different years with the
courses it has.

Todo:
 - listen to add/remove
 - move a course from sidebar to a year
 - move a course from one year to another
 - add/remove a year
 - display year in the title
 - display the total ects over all years
 - hide/show a year