YearCtrl now listens on course added and removed
When a course is added the course is added to the latest year.
When a course is removed it is removed from the year as well.

Added sort to the dragendsource. This allows you to differentiate
between a sidebar, one of the year view or one of the category.

You can now drag n drop from the sidebar to a year and from one year to
another and from yeat to sidebar