Simplified the fetching process
First each ctrl was responsible for fetching the right data at the right
time (before courses after faculties for example)

But that was getting over complicated. So now we going a little bit to
the redux side where we fetch all the necessary data to render the page
in the right order. So Storage is responsible for initialising all the
controllers and fetching the data.